Moved Ledger into it's own directory.
